🚚 move cache outside of folder
