+++ src/app/(app)/[user]/[repository]/issues/[number]/page.tsx
@@ -3,7 +3,7 @@ import "server-only";
 // components
 import { MarkdownTitle } from "~/components/markdown/markdown-title";
 import { Markdown } from "~/components/markdown/markdown";
-import { Cache } from "~/components/cache/cache";
+import { Cache } from "~/components/cache";
 
 // utils
 import { notFound } from "next/navigation";
